buy internet package for mci and payment page and opening redirect url in browser finished deep linking and redirect to desired page implementation is the next step

+++ iGap/General/Structs/IGStructPayment.swift
@@ -11,7 +11,7 @@ import Foundation
 struct IGStructPayment: Decodable {
     struct Info: Decodable {
         var product: Product?
-        var price: Float?
+        var price: Int?
         var vendor: String?
         var orderId: String?
